% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{draw_key_polygon_size}
\alias{draw_key_polygon_size}
\title{A polygon key glyphs for legends of size aesthetics}
\usage{
draw_key_polygon_size(width = 0.8, height = 0.8)
}
\arguments{
\item{width}{A numeric to control the width of keys,
which should be between 0 and 1. The default value is 0.8.}

\item{height}{A numeric to control the height of keys,
which should be between 0 and 1. The default value is 0.8.}
}
\value{
A Key glyphs function.
}
\description{
A polygon key glyphs for legends of size aesthetics
}
\examples{
library(dplyr)
library(sf)
library(ggplot2)


nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"),
                 quiet = TRUE)

nc2 <- nc \%>\% arrange(desc(AREA)) \%>\% slice(1:2)


ggplot(nc2) +
    geom_sf(aes(size = AREA),
            colour = "red") +
    scale_size(breaks = c(0.24, 0.241)) +
    theme(legend.key.size = unit(0.15, "npc"))


ggplot(nc2) +
    geom_sf(aes(size = AREA),
            colour = "red",
           key_glyph = draw_key_polygon_size()) +
    scale_size(breaks = c(0.24, 0.241)) +
    theme(legend.key.size = unit(0.15, "npc"))


ggplot(nc2) +
    geom_sf(aes(size = AREA),
            colour = "red",
            key_glyph = draw_key_polygon_size(0.9, 0.3)) +
    scale_size(breaks = c(0.24, 0.241)) +
    theme(legend.key.size = unit(0.15, "npc"))


}
