% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkey.R
\docType{package}
\name{ggkey}
\alias{ggkey}
\alias{ggkey-package}
\title{ggkey: some Key glyphs functions of ggplot2 to
   support more control of the keys}
\description{
ggkey: some Key glyphs functions of ggplot2 to
   support more control of the keys
}
